finance_project/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg==3.2.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
